Move useState to hooks.dart file
