+++ example/test/function_component_test.dart
@@ -1,5 +1,6 @@
 import 'dart:html';
 
+import 'package:react/hooks.dart';
 import 'package:react/react.dart' as react;
 import 'package:react/react_dom.dart' as react_dom;
 import 'package:react/react_client.dart';
@@ -7,7 +8,7 @@ import 'package:react/react_client.dart';
 var useStateTestFunctionComponent = react.registerFunctionComponent(UseStateTestComponent, displayName: 'useStateTest');
 
 UseStateTestComponent(Map props) {
-  final count = react.useState(0);
+  final count = useState(0);
 
   return react.div({}, [
     count.value,
